+++ app/database/models/users.py
@@ -64,8 +64,3 @@ class User(Base):
         """
         event.listen(cls, 'before_insert', cls.__set_user_role)
 
-    def is_admin(self):
-        return self.role == UserRole.admin
-
-    def is_moderator(self):
-        return self.role == UserRole.moderator
